% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Standardization.R
\name{convertSiToUs}
\alias{convertSiToUs}
\title{Convert SI unit value to US unit value}
\usage{
convertSiToUs(nameFromListOfMeasurements, siValue)
}
\arguments{
\item{nameFromListOfMeasurements}{one name of the measurement from list}

\item{siValue}{value to convert to US value}
}
\value{
data.frame with 4 columns: nameFromListOfMeasurements, siValue, siUnit, usValue, usUnit
}
\description{
Function converts SI values to US
}
