% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Standardization.R
\name{convertUsToSi}
\alias{convertUsToSi}
\title{Convert US unit value to SI unit value}
\usage{
convertUsToSi(nameFromListOfMeasurements, usValue)
}
\arguments{
\item{nameFromListOfMeasurements}{one name of the measurement from list}

\item{usValue}{value to convert to SI value}
}
\value{
data.frame with 4 columns: nameFromListOfMeasurements, siValue, siUnit, usValue, usUnit
}
\description{
Function converts US values to SI
}
