% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OrestCohortUtilits.R
\name{createCohortWithOrestUtilits}
\alias{createCohortWithOrestUtilits}
\title{placeTmpMeasurementMappedTable}
\usage{
createCohortWithOrestUtilits(
  connectionDetails,
  cdmSchema,
  orestTmpSchema,
  sqlCohortString,
  orestTmpMeasurementTable
)
}
\arguments{
\item{connectionDetails}{DatabaseConnector::createConnectionDetails S3 object}

\item{cdmSchema}{cdm database schema}

\item{orestTmpSchema}{schema to write}

\item{measurementName}{optional parameter prefix name of tmp measurement talbe}
}
\description{
This function will create a table with measurements that did not have a unit of measurement and will map source value using the Odysseus Data Services mapping date
}
