% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ckdEpi.R
\name{eGlomerularFiltrationRateCkdEpi}
\alias{eGlomerularFiltrationRateCkdEpi}
\title{CKD-EPI}
\usage{
eGlomerularFiltrationRateCkdEpi(creatinine, age, gender, race, creatinineUnit)
}
\arguments{
\item{creatinine}{Numeric vector. Serum creatinine, could be expressed in "umol/l", "mmol/l" or "mg/dL". Units of measurement should be defined in variable creatinineUnit (if not defined explicitly by user, the default value is "micromol/L").}

\item{age}{Numeric vector. Age, in years.}

\item{gender}{Vector}

\item{race}{Vector. race, specify in case of black patients. The value of variable refers to the parameter label_black.}

\item{creatinineUnit}{Character string. Units in which serum creatinne is expressed. Could be one of the following: "micromol/l", "mmol/l" or "mg/dl".}
}
\value{
numeric eGFR expressed in ml/min/1.73m<sup>2</sup>.
}
\description{
CKD-EPI
}
\details{
Calculate estimated glomerular filtration rate (eGFR) by CKD-EPI equation
Reference to the equation: Levey AS, Stevens LA, Schmid CH et al. A New Equation to Estimate Glomerular Filtration Rate. Ann Intern Med 2009;150:604–12.
}
\examples{
estimatedGlomerularFiltrationRateCkdEpi (creatinine = 1.4, age = 60, gender = "Male", race = "White",
  creatinineUnit = "mg/dl")
}
