% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReplaceMeasurements.R
\name{getCohortDetailsOrest}
\alias{getCohortDetailsOrest}
\title{Get Orest Cohort Details  (sql with replaces measurement info, concept ids)}
\usage{
getCohortDetailsOrest(
  sqlCohortString,
  orestTmpSchema,
  orestTmpMeasurementTable
)
}
\arguments{
\item{sqlCohortString}{sql cohort where there is value_as_number measurement criteia}

\item{orestTmpSchema}{schema to put Orest temp measurement table}

\item{orestTmpMeasurementTable}{measurement Orest temp table
Prepare sql and concept ids to use Orest measurement temp table}
}
\description{
getCohortDetailsOrest() prepares new sql cohort and concept ids from Atlas or Capr sql
}
