% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UnitsMapping2.R
\name{placeTmpMeasurementMappedTable}
\alias{placeTmpMeasurementMappedTable}
\title{placeTmpMeasurementMappedTable}
\usage{
placeTmpMeasurementMappedTable(
  measurementName = "any",
  connectionDetails,
  cdmSchema,
  writeSchema,
  targetMeasurementConceptIds
)
}
\arguments{
\item{measurementName}{optional parameter prefix name of tmp measurement talbe}

\item{connectionDetails}{DatabaseConnector::createConnectionDetails S3 object}

\item{cdmSchema}{cdm database schema}

\item{writeSchema}{schema to write}

\item{targetMeasurementConceptIds}{measurement concept ids to map}
}
\description{
This function will create a table with measurements that did not have a unit of measurement and will map source value using the Odysseus Data Services mapping date
}
