% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MappingPipeline.R
\name{runMappingPipeline}
\alias{runMappingPipeline}
\title{runMappingPipeline}
\usage{
runMappingPipeline(tbl, pathToOrest, source_code_field = TRUE)
}
\arguments{
\item{tbl}{data.frame with units source_code column}

\item{pathToOrest}{string absolute path to Orest package}
}
\value{
data.frame with 2 columns: source_code, concept_name
}
\description{
Function obtain tbl with source_code and provide 2 steps mapping
}
